+++ packages/firestore/test/unit/specs/spec_builder.ts
@@ -17,6 +17,7 @@
 
 import { IndexConfiguration } from '../../../src/api/index_configuration';
 import { ExpUserDataWriter } from '../../../src/api/reference_impl';
+import { FieldFilter, Filter } from '../../../src/core/filter';
 import {
   LimitType,
   newQueryForPath,
@@ -24,13 +25,7 @@ import {
   queryEquals,
   queryToTarget
 } from '../../../src/core/query';
-import {
-  canonifyTarget,
-  FieldFilter,
-  Filter,
-  Target,
-  targetEquals
-} from '../../../src/core/target';
+import { canonifyTarget, Target, targetEquals } from '../../../src/core/target';
 import { TargetIdGenerator } from '../../../src/core/target_id_generator';
 import { TargetId } from '../../../src/core/types';
 import { Document } from '../../../src/model/document';
@@ -1061,6 +1056,7 @@ export class SpecBuilder {
       return {
         key: SpecBuilder.keyToSpec(doc.key),
         version: doc.version.toMicroseconds(),
+        createTime: doc.createTime.toMicroseconds(),
         value: userDataWriter.convertValue(
           doc.data.value
         ) as JsonObject<unknown>,
@@ -1073,6 +1069,7 @@ export class SpecBuilder {
       return {
         key: SpecBuilder.keyToSpec(doc.key),
         version: doc.version.toMicroseconds(),
+        createTime: doc.createTime.toMicroseconds(),
         value: null
       };
     }