+++ packages/performance-compat/test/util.ts
@@ -31,12 +31,14 @@ export function getFakeApp(): FirebaseApp {
       appId: '1:777777777777:web:d93b5ca1475efe57'
     },
     automaticDataCollectionEnabled: true,
+    performance: () => ({} as any),
     delete: async () => {}
   };
 }
 
 export function getFakeModularPerformance(): FirebasePerformance {
   return {
+    app: getFakeApp(),
     instrumentationEnabled: true,
     dataCollectionEnabled: true
   };