+++ packages/util/src/environment.ts
@@ -140,7 +140,11 @@ export function isSafari(): boolean {
  * @return true if indexedDB is supported by current browser/service worker context
  */
 export function isIndexedDBAvailable(): boolean {
-  return typeof indexedDB === 'object';
+  try {
+    return typeof indexedDB === 'object';
+  } catch (e) {
+    return false;
+  }
 }
 
 /**